+++ tests/contrib/operators/test_databricks_operator.py
@@ -24,12 +24,12 @@ from datetime import datetime
 import mock
 
 import airflow.contrib.operators.databricks_operator as databricks_operator
-from airflow.contrib.hooks.databricks_hook import RunState
 from airflow.contrib.operators.databricks_operator import (
     DatabricksRunNowOperator, DatabricksSubmitRunOperator,
 )
 from airflow.exceptions import AirflowException
 from airflow.models import DAG
+from airflow.providers.databricks.hooks.databricks import RunState
 
 DATE = '2017-04-20'
 TASK_ID = 'databricks-operator'
