+++ tests/providers/databricks/hooks/__init__.py
@@ -16,18 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.databricks.hooks.databricks`."""
-
-import warnings
-
-# pylint: disable=unused-import
-from airflow.providers.databricks.hooks.databricks import (  # noqa
-    CANCEL_RUN_ENDPOINT, GET_RUN_ENDPOINT, RESTART_CLUSTER_ENDPOINT, RUN_LIFE_CYCLE_STATES, RUN_NOW_ENDPOINT,
-    START_CLUSTER_ENDPOINT, SUBMIT_RUN_ENDPOINT, TERMINATE_CLUSTER_ENDPOINT, USER_AGENT_HEADER,
-    DatabricksHook, RunState,
-)
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.databricks.hooks.databricks`.",
-    DeprecationWarning, stacklevel=2
-)
